Switch from BigInteger to Integer

+++ trunk/src/main/java/uk/ac/ebi/jmzml/model/mzml/Chromatogram.java
@@ -6,7 +6,6 @@ import uk.ac.ebi.jmzml.xml.jaxb.adapters.IdRefAdapter;
 import javax.xml.bind.annotation.*;
 import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 import java.io.Serializable;
-import java.math.BigInteger;
 
 
 /**
@@ -60,7 +59,7 @@ public class Chromatogram
 
     @XmlAttribute(required = true)
     @XmlSchemaType(name = "nonNegativeInteger")
-    protected BigInteger index;
+    protected Integer index;
 
     @XmlAttribute(required = true)
     protected int defaultArrayLength;
@@ -174,10 +173,10 @@ public class Chromatogram
      * 
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public BigInteger getIndex() {
+    public Integer getIndex() {
         return index;
     }
 
@@ -186,10 +185,10 @@ public class Chromatogram
      * 
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public void setIndex(BigInteger value) {
+    public void setIndex(Integer value) {
         this.index = value;
     }
