+++ trunk/src/main/java/uk/ac/ebi/jmzml/model/mzml/ComponentList.java
@@ -2,7 +2,6 @@ package uk.ac.ebi.jmzml.model.mzml;
 
 import javax.xml.bind.annotation.*;
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -46,7 +45,7 @@ public class ComponentList
     protected List<Component> components;
     @XmlAttribute(required = true)
     @XmlSchemaType(name = "nonNegativeInteger")
-    protected BigInteger count;
+    protected Integer count;
 
     /**
      * Gets the value of the components property.
@@ -81,9 +80,9 @@ public class ComponentList
      * Gets the value of the count property.
      *
      * @return possible object is
-     *         {@link BigInteger }
+     *         {@link Integer }
      */
-    public BigInteger getCount() {
+    public Integer getCount() {
         return count;
     }
 
@@ -91,9 +90,9 @@ public class ComponentList
      * Sets the value of the count property.
      *
      * @param value allowed object is
-     *              {@link BigInteger }
+     *              {@link Integer }
      */
-    public void setCount(BigInteger value) {
+    public void setCount(Integer value) {
         this.count = value;
     }
