+++ trunk/src/main/java/uk/ac/ebi/jmzml/model/mzml/DataProcessingList.java
@@ -1,16 +1,10 @@
 
 package uk.ac.ebi.jmzml.model.mzml;
 
+import javax.xml.bind.annotation.*;
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlAttribute;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlSchemaType;
-import javax.xml.bind.annotation.XmlType;
 
 
 /**
@@ -49,7 +43,7 @@ public class DataProcessingList
     protected List<DataProcessing> dataProcessing;
     @XmlAttribute(required = true)
     @XmlSchemaType(name = "nonNegativeInteger")
-    protected BigInteger count;
+    protected Integer count;
 
     /**
      * Gets the value of the dataProcessing property.
@@ -85,10 +79,10 @@ public class DataProcessingList
      * 
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public BigInteger getCount() {
+    public Integer getCount() {
         return count;
     }
 
@@ -97,10 +91,10 @@ public class DataProcessingList
      * 
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@link Integer }
      *     
      */
-    public void setCount(BigInteger value) {
+    public void setCount(Integer value) {
         this.count = value;
     }
